refactor(types/sync): rename files to PascalCase, add index
