+++ src/types/sync/AuthStrategy.ts
@@ -1,0 +1,4 @@
+/**
+ * Per-endpoint authentication strategy
+ */
+export type AuthStrategy = "none" | "bearer" | "basic" | "custom";
