+++ src/types/sync/ConflictStrategy.ts
@@ -1,0 +1,8 @@
+/**
+ * Conflict resolution strategy
+ */
+export type ConflictStrategy =
+  | "lastWriteWins"
+  | "serverWins"
+  | "clientWins"
+  | "manual";
