+++ src/types/sync/IAuthenticationDefinition.ts
@@ -1,0 +1,10 @@
+import type { AuthStrategy } from "./AuthStrategy";
+
+/**
+ * Per-endpoint authentication.
+ */
+export interface IAuthenticationDefinition {
+  strategy: AuthStrategy;
+  tokenKey?: string;
+  headerName?: string;
+}
