+++ src/types/sync/IBackgroundDefinition.ts
@@ -1,10 +1,7 @@
-/** Background sync configuration for an {@link ISyncDefinition}. */
+/** Background sync configuration */
 export interface IBackgroundDefinition {
   enabled: boolean;
-
   minimumInterval: number;
-
   requiresNetwork?: boolean;
-
   requiresCharging?: boolean;
 }
