+++ src/types/sync/INativeSyncResult.ts
@@ -1,14 +1,9 @@
 /** Result of a sync run returned by the Native Sync Engine. */
 export interface INativeSyncResult {
   cursor?: string;
-
   operationsApplied: number;
-
   inserted: number;
-
   updated: number;
-
   deleted: number;
-
   duration: number;
 }
