+++ src/types/sync/IRequestDefinition.ts
@@ -1,4 +1,4 @@
-import type { RequestExpression } from "./request-expression";
+import type { RequestExpression } from "./RequestExpression";
 
 /** Declarative body of a sync request. */
 export interface IRequestDefinition<_TEntity> {
