+++ src/types/sync/ISyncOperation.ts
@@ -1,12 +1,8 @@
 /** Individual operation in the sync queue (`sync_queue`), generated by a trigger. */
 export interface ISyncOperation {
   operation: "insert" | "update" | "delete";
-
   entity: string;
-
   primaryKey: string;
-
   payload: Record<string, unknown>;
-
   updatedAt: number;
 }
