+++ src/types/sync/ITransport.ts
@@ -1,0 +1,7 @@
+/**
+ * Transport protocol.
+ */
+export type Transport =
+  | "rest"
+  | "graphql"
+  | "grpc";
