+++ src/types/sync/SyncDirection.ts
@@ -1,0 +1,7 @@
+/**
+ * Sync direction
+ */
+export type SyncDirection =
+  | "bidirectional"
+  | "push"
+  | "pull";
