+++ src/types/sync/SyncStrategy.ts
@@ -1,0 +1,7 @@
+/**
+ * Sync strategy
+ */
+export type SyncStrategy =
+  | "operations"
+  | "incremental"
+  | "full";
