--- src/types/sync/auth-strategy.ts
@@ -1,6 +1,0 @@
-/**
- * Per-endpoint authentication strategy for an {@link IAuthenticationDefinition}.
- *
- * Not to be confused with {@link AuthProvider}, which is the global app credential.
- */
-export type AuthStrategy = "none" | "bearer" | "basic" | "custom";
