--- src/types/sync/compression-definition.ts
@@ -1,6 +1,0 @@
-/** Out of MVP scope. Not referenced in any `ISyncDefinition` yet. */
-export interface ICompressionDefinition {
-  enabled: boolean;
-
-  algorithm: "gzip" | "brotli";
-}
