--- src/types/sync/conflict-strategy.ts
@@ -1,10 +1,0 @@
-/**
- * Conflict resolution strategy for an {@link ISyncDefinition}.
- *
- * MVP: only `"lastWriteWins"` is implemented.
- */
-export type ConflictStrategy =
-  | "lastWriteWins" // MVP: only one implemented
-  | "serverWins"
-  | "clientWins"
-  | "manual";
