--- src/types/sync/encryption-definition.ts
@@ -1,6 +1,0 @@
-/** Out of MVP scope. Not referenced in any `ISyncDefinition` yet. */
-export interface IEncryptionDefinition {
-  enabled: boolean;
-
-  algorithm: "aes256";
-}
